// src/app/core/Backend/model/empresa.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Direccion } from './direccion';

export interface Empresa { 
    direccion?: Direccion;
    nombre?: string;
    personaId?: string;
    ruc?: string;
}